requirements.txt
flask
gunicorn
openai
python-dotenv
requests
beautifulsoup4
pypdf